pyproject.toml
[project]
name = "deep-research-gemini"
version = "0.1.0"
description = "A deep research tool powered by Gemini and AutoGen"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit==1.43.0",
    # "pyautogen==0.8.2",
    "google-generativeai==0.8.4",
    "python-dotenv==1.0.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "google-genai==1.7.0"
]
